onyx: Add syscall_ret hack for older Onyx versions
...which don't do proper errno casting.

+++ src/internal/syscall_ret.c
@@ -3,7 +3,8 @@
 
 long __syscall_ret(unsigned long r)
 {
-	if (r > -4096UL) {
+	/* TODO: We need this cast because the kernel doesn't cast return values properly, just yet... */
+	if ((int) r > -4096UL) {
 		errno = -r;
 		return -1;
 	}
